requirements.txt
# Deep Learning Framework
torch>=1.10.0
torchvision>=0.11.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0

# Image Processing
Pillow>=8.0.0

# Visualization
matplotlib>=3.4.0

# Data Download
opendatasets>=0.1.22

# Web Service
Flask>=2.3.0
gunicorn>=21.0.0
requests>=2.28.0

# Progress bars
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
